+++ src/wallet/wallet_ismine.h
@@ -1,0 +1,31 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2015 The Bitcoin Core developers
+// Copyright (c) 2016-2016 Silk Network
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef DARKSILK_WALLET_WALLET_ISMINE_H
+#define DARKSILK_WALLET_WALLET_ISMINE_H
+
+#include "script/script.h"
+
+#include <stdint.h>
+
+class CKeyStore;
+class CScript;
+
+/// IsMine() return codes
+enum isminetype
+{
+    ISMINE_NO = 0,
+    ISMINE_WATCH_ONLY = 1,
+    ISMINE_SPENDABLE = 2,
+    ISMINE_ALL = ISMINE_WATCH_ONLY | ISMINE_SPENDABLE
+};
+/// used for bitflags of isminetype
+typedef uint8_t isminefilter;
+
+isminetype IsMine(const CKeyStore& keystore, const CScript& scriptPubKey);
+isminetype IsMine(const CKeyStore& keystore, const CTxDestination& dest);
+
+#endif // DARKSILK_WALLET_WALLET_ISMINE_H
